reintroduces openai_chat_completion_base
