+++ python/semantic_kernel/connectors/ai/open_ai/services/open_ai_config_base.py
@@ -101,8 +101,3 @@ class OpenAIConfigBase(OpenAIHandler):
         )
         base.update(client_settings)
         return base
-
-    def get_model_args(self) -> Dict[str, Any]:
-        return {
-            "model": self.ai_model_id,
-        }
