+++ python/semantic_kernel/connectors/ai/open_ai/services/open_ai_handler.py
@@ -111,72 +111,3 @@ class OpenAIHandler(AIServiceClientBase, ABC):
     def get_request_settings_class(self) -> "AIRequestSettings":
         """Return the class with the applicable request settings."""
         return OpenAIRequestSettings
-
-    def get_metadata_from_chat_response(self, response: ChatCompletion) -> Dict[str, Any]:
-        return {
-            "id": response.id,
-            "created": response.created,
-            "system_fingerprint": response.system_fingerprint,
-            "usage": response.usage,
-        }
-
-    def get_metadata_from_streaming_chat_response(self, response: ChatCompletionChunk) -> Dict[str, Any]:
-        return {
-            "id": response.id,
-            "created": response.created,
-            "system_fingerprint": response.system_fingerprint,
-        }
-
-    def get_metadata_from_text_response(self, response: Completion) -> Dict[str, Any]:
-        return {
-            "id": response.id,
-            "created": response.created,
-            "system_fingerprint": response.system_fingerprint,
-            "usage": response.usage,
-        }
-
-    def get_metadata_from_streaming_text_response(self, response: Completion) -> Dict[str, Any]:
-        return {
-            "id": response.id,
-            "created": response.created,
-            "system_fingerprint": response.system_fingerprint,
-        }
-
-    def get_metadata_from_chat_choice(self, choice: Union[Choice, ChunkChoice]) -> Dict[str, Any]:
-        return {
-            "finish_reason": choice.finish_reason,
-            "index": choice.index,
-            "logprobs": choice.logprobs,
-        }
-
-    def get_tool_calls_from_chat_choice(self, choice: Union[Choice, ChunkChoice]) -> Optional[List[ToolCall]]:
-        if isinstance(choice, Choice):
-            content = choice.message
-        else:
-            content = choice.delta
-        if content.tool_calls is None:
-            return None
-        return [
-            ToolCall(
-                id=tool.id,
-                type=tool.type,
-                function=FunctionCall(name=tool.function.name, arguments=tool.function.arguments),
-            )
-            for tool in content.tool_calls
-        ]
-
-    def get_function_call_from_chat_choice(self, choice: Union[Choice, ChunkChoice]) -> Optional[FunctionCall]:
-        if isinstance(choice, Choice):
-            content = choice.message
-        else:
-            content = choice.delta
-        if content.function_call is None:
-            return None
-        return FunctionCall(name=content.function_call.name, arguments=content.function_call.arguments)
-
-    def get_metadata_from_text_choice(self, choice: CompletionChoice) -> Dict[str, Any]:
-        return {
-            "finish_reason": choice.finish_reason,
-            "index": choice.index,
-            "logprobs": choice.logprobs,
-        }
